feat: override fish's 'man' implementation
When accessing the man page of some C functions like 'stat' or 'exec', I
was always met with the man page for the equivalent fish function. I
dislike having to always add a 2 or a 3 for the section.  Another reason
why I went about it this way was to use 'nvim' as my MANPAGER. I wasn't
able to just use MANPAGER since man doesn't support some shell constructs
in the variable, particulary pipes in my case.

If I want to access the man page for fish's functions, I will use
'fman'. 'fman' is short for fish's man. This way I can also use
'nvim' as my MANPAGER.

One of the coolest things to come out of this is that man pages
follow the color scheme I have set in 'nvim'. Colored man pages
is the second reason for me going this route. The first being
bypassing fish's man pages.