make the length of `XDG_RUNTIME_DIR` smaller
`podman` complains that it can't connect to any socket when trying to
drop into a container. Looking on different forums, it seems to be
because of the length of the path to the socket. The previous
`XDG_RUNTIME_DIR` was a bit lengthy and probably overflow'd the buffer
where the path is stored. Reducing the length of `XDG_RUNTIME_DIR` fixes
this issue.