rename passert to assert_ptr
I want to be able to type `assert` and have my completion have both
`assert` and `assert_ptr` come up. Before, only one of them would
come up due to the sorting of the list.

Also, add an `assert` snippet to test for positive numbers with the
same naming scheme as `assert_ptr`.