import `jony_ng`
This colorscheme version supports truecolors. It is aware of the current
base 16 colors of the terminal. The base 16 colors in this file have to
be kept in sync with the terminal emulator's config. Otherwise, there
may be some regressions.

I wrote some functions that help wrap the logic of deducing when to use
the changed base 16 colors. Going over it again now, it seems
over-engineered... It makes it easier to read what's going on though, as
in, knowing what values are set to what instead of a mis-mash of the
highlight arguments.

These functions will also make it easier to wrap the handling of only
'gui' colors and calculating the approximate rgb value of the 'cterm'
colors if there isn't support for 'termguicolors'.