nvim/config: refactor entire config into lua modules
All of the nvim specific configuration has been split up into lua
modules. The following items were split up:

1. lsp
2. lsp_signature
3. treesitter
4. cmp (and its completion sources)

Each module defines a 'setup()' method that, well, setups each module.
This is so that the orders of the 'require()' don't influence the final
configuration.

Most of the work is complete. The only thing left is to enforce a
consistent style on all lua files. This can be done using a lua
formatter which I have yet to setup. I will add that in a separate
commit.

I also took the liberty of removing my 'compe' config since it's in
maintenance mode only.