nvim/ulti_snips/c: use `++i` instead of `i++` in `for` snippet
I know compilers might be able to tell that only the increment, and not
the copy is needed but, from what I've read online, it might matter for
cpp. Regardless, '++i' makes more sense anyways since the temporary copy
isn't used in the third part of the 'for' loop so it doesn't hurt.