nvim/lua/gitsigns.lua: switch to `gitsigns` over `gitgutter`
For some reason, the highlighting that was supposed to show in hunk
previews, never did. I tried to fix it but I couldn't get it to work. I
had been thinking about switching to 'gitsigns' anyways because of the
line blame functionality. I might have to mess with the highlighting a
bit to get a scheme I like but at least diff highlighting works how I
want it to in 'gitsigns'.